+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # `multipart-parser` CHANGELOG
 
+## HEAD
+
+- Provide correct type for `part.arrayBuffer()`
+- `part.isFile` now defaults to `true` if `part.mediaType === 'application/octet-stream'`
+
 ## v0.6.1 (Aug 18, 2024)
 
 - More small performance improvements